[New] Added test cases for record list components date column
